feat: version negotiation (#639)
# Description

Adding version negotiation between `zipper` and `source/sfn`, if version
negotiation failed, the `source/sfn` cannot connect to `zipper`.

The version format must follow the `Major.Minor.Patch` formatting, and
the Major, Minor, and Patch components must be numeric.

The client‘s MAJOR and MINOR versions should equal to server's,
otherwise, the zipper will be considered has break-change, the handshake
will fail.


## Impact

This is a break-change. It will not be possible to establish a
connection with the `zipper` if the `source/sfn` does not send the
version field.