feat: cli doc (#1115)
This pull request introduces functionality to generate CLI documentation
in Markdown format, adds an example template for running a zipper, and
updates dependencies to support the new features. The most important
changes include the addition of the `Doc` function for generating
documentation, the integration of the zipper example template, and
corresponding updates to dependencies and tests.